Extract AssertTimeoutPreemptively from AssertTimeout
